Adding initial skill bar implementation
Still needs work. The skill bar needs to be changed to work with each
skill's icon and populate the skill bar with the skills the player has.
The optional toggle system is still not implemented.
The config/settings/keybindings need to be reorganized to allow for the
use of the methods in the proxy classes.