-implements ISubSkillTypeIdentifier
Note; I've not made SLB extend Skills just yet, because
ISubSkillTypeIdentifier doesn't provide anything. Causing all skills to
have that in their constructor when it's null seems pointless - and
currently all skill stats can be obtained anywhere. When it has a use,
I'll use it :)

Skills do implement it though, so it's ready on that front.