Working Dimension Teleport
Teleporting to the dimension on login is now working, the dimension is
properly created, with the correct parameters and fields, and is now
creating some of the correct world gen. However, the original "giant
circle wall" is being eaten by vanilla biomes, so that needs to be fixed
still so that everything but aincrad is void.