Fixed Dungeon Generation
Dungeon generation is now once again functional, outside of the bounds
of a 0,0 origin point. It now will generate a dungeon outside of 300
meters away from the center of the floor (or some other configurable
value in the future) and within the size of the dungeon distance away
from the wall, randomly, as many dungeons as we want, and as many pieces
as we want, including whichever schematics we choose to use. The
foundation is laid, now to add more schema!