Added SAOZone Class
I created a new cuboid class, that includes the origin point, which allows us to more easily math out zone calculations. SAOZone also has some basic functions which can be expanded upon. We need to start working on the entity and "action" classes in order to add more functionality to it.